1st version usable
